+++ components/user_profile.py
@@ -1,5 +1,5 @@
 import allure
-from playwright.sync_api import Page
+from playwright.sync_api import Page, expect
 
 
 class UserProfile:
@@ -12,3 +12,16 @@ class UserProfile:
                 self.page.url
                 == "http://tester:dslfjsdfblkhew%40122b1klbfw@testing.misleplav.ru/subscription/profile/"
         )
+
+    @allure.step("Проверяем наличие кнопки 'Выйти' в профиле студента")
+
+    def check_exit_btn_exists_for_student(self):
+        exit_button = self.page.get_by_role("link", name="Выйти из аккаунта")
+        return exit_button.is_visible()
+
+    @allure.step("Кликаем кнопку 'Выйти' в профиле студента")
+    def click_exit_btn_for_student(self):
+        button = self.page.locator("//a[@href= '/authorization/logout/']")
+        button.click()
+        expect(self.page).to_have_url(
+            "http://testing.misleplav.ru/")
